Move out Edge to edge.rs from routing.rs (#5284)
Code is going to be much cleaner once we move out `Edge` and related structs to a separate file.
