+++ chain/network/src/routing/routing.rs
@@ -9,9 +9,8 @@ use actix::{Actor, Message};
 use borsh::{BorshDeserialize, BorshSerialize};
 use cached::{Cached, SizedCache};
 use conqueue::{QueueReceiver, QueueSender};
-use near_crypto::{KeyType, SecretKey, Signature};
 #[cfg(feature = "test_features")]
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use tracing::warn;
 
 use near_primitives::hash::CryptoHash;
@@ -19,6 +18,7 @@ use near_primitives::network::{AnnounceAccount, PeerId};
 use near_primitives::types::AccountId;
 use near_store::{ColAccountAnnouncements, Store};
 
+use crate::routing::edge::{Edge, SimpleEdge};
 use crate::routing::route_back_cache::RouteBackCache;
 use crate::PeerInfo;
 use crate::{
@@ -41,294 +41,11 @@ pub const DELETE_PEERS_AFTER_TIME: Duration = Duration::from_secs(3_600);
 /// Graph implementation supports up to 128 peers.
 pub const MAX_NUM_PEERS: usize = 128;
 
-/// Information that will be ultimately used to create a new edge.
-/// It contains nonce proposed for the edge with signature from peer.
-#[derive(Clone, BorshSerialize, BorshDeserialize, PartialEq, Eq, Debug, Default)]
-pub struct EdgeInfo {
-    pub nonce: u64,
-    pub signature: Signature,
-}
-
-impl EdgeInfo {
-    pub fn new(peer0: &PeerId, peer1: &PeerId, nonce: u64, secret_key: &SecretKey) -> Self {
-        let data = if peer0 < peer1 {
-            EdgeInner::build_hash(&peer0, &peer1, nonce)
-        } else {
-            EdgeInner::build_hash(&peer1, &peer0, nonce)
-        };
-
-        let signature = secret_key.sign(data.as_ref());
-        Self { nonce, signature }
-    }
-}
-
 /// Status of the edge
 #[derive(BorshSerialize, BorshDeserialize, Clone, PartialEq, Eq, Debug, Hash)]
 pub enum EdgeType {
     Added,
     Removed,
-}
-
-#[derive(BorshSerialize, BorshDeserialize, Clone, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "test_features", derive(Serialize, Deserialize))]
-pub struct Edge(pub Arc<EdgeInner>);
-
-impl Edge {
-    /// Create an addition edge.
-    pub fn new(
-        peer0: PeerId,
-        peer1: PeerId,
-        nonce: u64,
-        signature0: Signature,
-        signature1: Signature,
-    ) -> Self {
-        Edge(Arc::new(EdgeInner::new(peer0, peer1, nonce, signature0, signature1)))
-    }
-
-    pub fn make_fake_edge(peer0: PeerId, peer1: PeerId, nonce: u64) -> Self {
-        Self(Arc::new(EdgeInner {
-            key: (peer0, peer1),
-            nonce,
-            signature0: Signature::empty(KeyType::ED25519),
-            signature1: Signature::empty(KeyType::ED25519),
-            removal_info: None,
-        }))
-    }
-
-    /// Build a new edge with given information from the other party.
-    pub fn build_with_secret_key(
-        peer0: PeerId,
-        peer1: PeerId,
-        nonce: u64,
-        secret_key: &SecretKey,
-        signature1: Signature,
-    ) -> Self {
-        let hash = if peer0 < peer1 {
-            Self::build_hash(&peer0, &peer1, nonce)
-        } else {
-            Self::build_hash(&peer1, &peer0, nonce)
-        };
-        let signature0 = secret_key.sign(hash.as_ref());
-        Self::new(peer0, peer1, nonce, signature0, signature1)
-    }
-
-    /// Build the hash of the edge given its content.
-    /// It is important that peer0 < peer1 at this point.
-    pub fn build_hash(peer0: &PeerId, peer1: &PeerId, nonce: u64) -> CryptoHash {
-        CryptoHash::hash_borsh(&(peer0, peer1, nonce))
-    }
-
-    pub fn make_key(peer0: PeerId, peer1: PeerId) -> (PeerId, PeerId) {
-        if peer0 < peer1 {
-            (peer0, peer1)
-        } else {
-            (peer1, peer0)
-        }
-    }
-
-    /// Helper function when adding a new edge and we receive information from new potential peer
-    /// to verify the signature.
-    pub fn partial_verify(peer0: PeerId, peer1: PeerId, edge_info: &EdgeInfo) -> bool {
-        let pk = peer1.public_key();
-        let data = if peer0 < peer1 {
-            Edge::build_hash(&peer0, &peer1, edge_info.nonce)
-        } else {
-            Edge::build_hash(&peer1, &peer0, edge_info.nonce)
-        };
-        edge_info.signature.verify(data.as_ref(), &pk)
-    }
-
-    /// Next nonce of valid addition edge.
-    pub fn next_nonce(nonce: u64) -> u64 {
-        if nonce % 2 == 1 {
-            nonce + 2
-        } else {
-            nonce + 1
-        }
-    }
-}
-
-impl std::ops::Deref for Edge {
-    type Target = EdgeInner;
-
-    fn deref(&self) -> &EdgeInner {
-        &self.0
-    }
-}
-
-/// Edge object. Contains information relative to a new edge that is being added or removed
-/// from the network. This is the information that is required.
-#[derive(BorshSerialize, BorshDeserialize, Clone, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "test_features", derive(Serialize, Deserialize))]
-pub struct EdgeInner {
-    /// Since edges are not directed `key.0 < peer1` should hold.
-    pub key: (PeerId, PeerId),
-    /// Nonce to keep tracking of the last update on this edge.
-    /// It must be even
-    pub nonce: u64,
-    /// Signature from parties validating the edge. These are signature of the added edge.
-    pub signature0: Signature,
-    pub signature1: Signature,
-    /// Info necessary to declare an edge as removed.
-    /// The bool says which party is removing the edge: false for Peer0, true for Peer1
-    /// The signature from the party removing the edge.
-    pub removal_info: Option<(bool, Signature)>,
-}
-
-impl EdgeInner {
-    /// Create an addition edge.
-    pub fn new(
-        peer0: PeerId,
-        peer1: PeerId,
-        nonce: u64,
-        signature0: Signature,
-        signature1: Signature,
-    ) -> Self {
-        let (peer0, signature0, peer1, signature1) = if peer0 < peer1 {
-            (peer0, signature0, peer1, signature1)
-        } else {
-            (peer1, signature1, peer0, signature0)
-        };
-
-        Self { key: (peer0, peer1), nonce, signature0, signature1, removal_info: None }
-    }
-
-    pub fn key(&self) -> &(PeerId, PeerId) {
-        &self.key
-    }
-
-    pub fn to_simple_edge(&self) -> SimpleEdge {
-        SimpleEdge::new(self.key.0.clone(), self.key.1.clone(), self.nonce)
-    }
-
-    /// Create the remove edge change from an added edge change.
-    pub fn remove_edge(&self, my_peer_id: PeerId, sk: &SecretKey) -> Edge {
-        assert_eq!(self.edge_type(), EdgeType::Added);
-        let mut edge = self.clone();
-        edge.nonce += 1;
-        let me = edge.key.0 == my_peer_id;
-        let hash = edge.hash();
-        let signature = sk.sign(hash.as_ref());
-        edge.removal_info = Some((me, signature));
-        Edge(Arc::new(edge))
-    }
-
-    /// Build the hash of the edge given its content.
-    /// It is important that peer0 < peer1 at this point.
-    pub fn build_hash(peer0: &PeerId, peer1: &PeerId, nonce: u64) -> CryptoHash {
-        debug_assert!(peer0 < peer1);
-        CryptoHash::hash_borsh(&(peer0, peer1, &nonce))
-    }
-
-    fn hash(&self) -> CryptoHash {
-        Edge::build_hash(&self.key.0, &self.key.1, self.nonce)
-    }
-
-    fn prev_hash(&self) -> CryptoHash {
-        Edge::build_hash(&self.key.0, &self.key.1, self.nonce - 1)
-    }
-
-    pub fn verify(&self) -> bool {
-        if self.key.0 > self.key.1 {
-            return false;
-        }
-
-        match self.edge_type() {
-            EdgeType::Added => {
-                let data = self.hash();
-
-                self.removal_info.is_none()
-                    && self.signature0.verify(data.as_ref(), &self.key.0.public_key())
-                    && self.signature1.verify(data.as_ref(), &self.key.1.public_key())
-            }
-            EdgeType::Removed => {
-                // nonce should be an even positive number
-                if self.nonce == 0 {
-                    return false;
-                }
-
-                // Check referring added edge is valid.
-                let add_hash = self.prev_hash();
-                if !self.signature0.verify(add_hash.as_ref(), &self.key.0.public_key())
-                    || !self.signature1.verify(add_hash.as_ref(), &self.key.1.public_key())
-                {
-                    return false;
-                }
-
-                if let Some((party, signature)) = &self.removal_info {
-                    let peer = if *party { &self.key.0 } else { &self.key.1 };
-                    let del_hash = self.hash();
-                    signature.verify(del_hash.as_ref(), &peer.public_key())
-                } else {
-                    false
-                }
-            }
-        }
-    }
-
-    pub fn get_pair(&self) -> &(PeerId, PeerId) {
-        &self.key
-    }
-
-    /// It will be considered as a new edge if the nonce is odd, otherwise it is canceling the
-    /// previous edge.
-    pub fn edge_type(&self) -> EdgeType {
-        if self.nonce % 2 == 1 {
-            EdgeType::Added
-        } else {
-            EdgeType::Removed
-        }
-    }
-    /// Next nonce of valid addition edge.
-    pub fn next(&self) -> u64 {
-        Edge::next_nonce(self.nonce)
-    }
-
-    pub fn contains_peer(&self, peer_id: &PeerId) -> bool {
-        self.key.0 == *peer_id || self.key.1 == *peer_id
-    }
-
-    /// Find a peer id in this edge different from `me`.
-    pub fn other(&self, me: &PeerId) -> Option<&PeerId> {
-        if self.key.0 == *me {
-            Some(&self.key.1)
-        } else if self.key.1 == *me {
-            Some(&self.key.0)
-        } else {
-            None
-        }
-    }
-}
-
-/// Represents edge between two nodes. Unlike `Edge` it doesn't contain signatures.
-#[derive(Hash, Clone, Eq, PartialEq, Debug)]
-#[cfg_attr(feature = "test_features", derive(Serialize, Deserialize))]
-pub struct SimpleEdge {
-    key: (PeerId, PeerId),
-    nonce: u64,
-}
-
-impl SimpleEdge {
-    pub fn new(peer0: PeerId, peer1: PeerId, nonce: u64) -> SimpleEdge {
-        let (peer0, peer1) = Edge::make_key(peer0, peer1);
-        SimpleEdge { key: (peer0, peer1), nonce }
-    }
-
-    pub fn key(&self) -> &(PeerId, PeerId) {
-        &self.key
-    }
-
-    pub fn nonce(&self) -> u64 {
-        self.nonce
-    }
-
-    pub fn edge_type(&self) -> EdgeType {
-        if self.nonce % 2 == 1 {
-            EdgeType::Added
-        } else {
-            EdgeType::Removed
-        }
-    }
 }
 
 #[derive(BorshSerialize, BorshDeserialize, PartialEq, Eq, Clone, Debug, Copy)]
