+++ chain/network/src/routing/routing_table_actor.rs
@@ -8,6 +8,8 @@ use actix::{Actor, Addr, Context, Handler, Message, System};
 use tracing::error;
 use tracing::{debug, trace, warn};
 
+#[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
+use crate::routing::edge::SimpleEdge;
 #[cfg(feature = "delay_detector")]
 use delay_detector::DelayDetector;
 use near_performance_metrics_macros::perf;
@@ -17,15 +19,16 @@ use near_primitives::utils::index_to_bytes;
 use near_store::db::DBCol::{ColComponentEdges, ColLastComponentNonce, ColPeerComponent};
 use near_store::{Store, StoreUpdate};
 
+use crate::routing::edge::Edge;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 use crate::routing::ibf::{Ibf, IbfBox};
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 use crate::routing::ibf_peer_set::IbfPeerSet;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 use crate::routing::ibf_set::IbfSet;
-use crate::routing::routing::{Edge, EdgeType, Graph, SAVE_PEERS_MAX_TIME};
+use crate::routing::routing::{EdgeType, Graph, SAVE_PEERS_MAX_TIME};
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
-use crate::routing::routing::{SimpleEdge, ValidIBFLevel, MIN_IBF_LEVEL};
+use crate::routing::routing::{ValidIBFLevel, MIN_IBF_LEVEL};
 use crate::stats::metrics;
 use crate::types::StopMsg;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
